# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Backend API",
        "description": "Backend API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Create User Account",
                "operationId": "942681186878cb5e89feacd1832acfe5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success Create Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegisterResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageWithErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Something error on the server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DefaultResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "authenticate account",
                "operationId": "4792d88bfaf2b1bca3ff19daafa8c1dc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequestSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "headers": {
                            "Authorization": {
                                "description": "JWT access token",
                                "schema": {
                                    "type": "string",
                                    "example": "Bearer eyJ..."
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DefaultResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageWithErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DefaultResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized (password or email not valid)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DefaultResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Something error on the server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DefaultResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "operationId": "a667dbdc4334fbe2df97006ad6f27ebd",
                "parameters": [
                    {
                        "name": "Authentication",
                        "in": "header",
                        "description": "Client jwt access token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success re-create token",
                        "headers": {
                            "Authorization": {
                                "description": "JWT access token",
                                "schema": {
                                    "type": "string",
                                    "example": "Bearer eyJ..."
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DefaultResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DefaultResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/createinfo": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "create user info",
                "operationId": "d25bb07b765a6eae87fe2f6dc8709bd8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserInfoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success Create User Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DefaultResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DefaultResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageWithErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreateUserInfoRequest": {
                "required": [
                    "sekolah",
                    "jurusan",
                    "nomor_telepon",
                    "periode_awal",
                    "periode_akhir"
                ],
                "properties": {
                    "account_id": {
                        "type": "integer"
                    },
                    "sekolah": {
                        "type": "string"
                    },
                    "jurusan": {
                        "type": "string"
                    },
                    "nomor_telepon": {
                        "type": "string"
                    },
                    "periode_awal": {
                        "type": "string",
                        "format": "date"
                    },
                    "periode_akhir": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
            "DefaultResponse": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "LoginRequestSchema": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "MessageWithErrorResponse": {
                "required": [
                    "message",
                    "error"
                ],
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "error": {
                        "type": "error_obj"
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "required": [
                    "username",
                    "email",
                    "password"
                ],
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "RegisterResponse": {
                "required": [
                    "message",
                    "account_id"
                ],
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "account_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "User",
            "description": "User"
        }
    ]
}